# docs/P3/03-api-specification.yaml
openapi: "3.0.3"
info:
  title: CampusHub API
  description: |
    校园互助服务平台 API 规范文档

    覆盖模块：认证与用户、需求与任务大厅、接单与订单、消息通知、评价与信用、举报、文件上传、后台管理
  version: "1.0.0"
  contact:
    name: 暴风星云裂
    url: https://gitee.com/duomaameng/campus-hub
servers:
  - url: http://localhost:8080/api
    description: 本地开发环境
  - url: https://campushub.nju.edu.cn/api
    description: 生产环境
...
security:
  - bearerAuth: []
paths:
...
  /users/me:
    get:
      tags: [User]
      summary: 获取当前用户资料
      responses:
        "200":
          description: 用户资料
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfileResponse"
    patch:
      tags: [User]
      summary: 修改个人资料
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileRequest"
      responses:
        "200":
          description: 更新后的用户资料
    delete:
      tags: [User]
      summary: 注销账号
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [password]
              properties:
                password:
                  type: string
  /users/{userId}/profile:
    get:
      tags: [User]
      summary: 获取用户公开资料
      parameters:
        - $ref: "#/components/parameters/UserIdPath"
      responses:
        "200":
          description: 公开资料
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicProfileResponse"
  /users/{userId}/reviews:
    get:
      tags: [User]
      summary: 获取用户评价记录
      parameters:
        - $ref: "#/components/parameters/UserIdPath"
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Size"
      responses:
        "200":
          description: 评价记录列表
  /users/{userId}/credit:
    get:
      tags: [User]
      summary: 获取用户信用信息
      parameters:
        - $ref: "#/components/parameters/UserIdPath"
      responses:
        "200":
          description: 信用信息
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreditInfoResponse"

  # ==================== Task ====================
...
  /orders:
    get:
      tags: [Order]
      summary: 我的订单列表
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Size"
        - name: role
          in: query
          schema:
            type: string
            enum: [PUBLISHER, PROVIDER]
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/OrderStatus"
        - name: keyword
          in: query
          schema:
            type: string
  /orders/{orderId}:
    get:
      tags: [Order]
      summary: 订单详情
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: integer
            format: int64
  /orders/{orderId}/complete:
    post:
      tags: [Order]
      summary: 服务方提交完成凭证
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [proofImageId]
              properties:
                proofImageId:
                  type: integer
                  format: int64
                note:
                  type: string
                  maxLength: 500
  /orders/{orderId}/confirm-completion:
    post:
      tags: [Order]
      summary: 发布者确认完成
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: integer
            format: int64
  /orders/{orderId}/cancel:
    post:
      tags: [Order]
      summary: 取消订单
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                reason:
                  type: string
...
  /orders/{orderId}/status-logs:
    get:
      tags: [Order]
      summary: 订单状态变更日志
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: integer
            format: int64

  # ==================== Notification ====================
  /notifications:
    get:
      tags: [Notification]
      summary: 通知列表
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Size"
        - name: type
          in: query
          schema:
            $ref: "#/components/schemas/NotificationType"
        - name: read
          in: query
          schema:
            type: boolean
...
  /notifications/{notificationId}/read:
    patch:
      tags: [Notification]
      summary: 标记通知为已读
      parameters:
        - name: notificationId
          in: path
          required: true
          schema:
            type: integer
            format: int64
  /notifications/{notificationId}:
    delete:
      tags: [Notification]
      summary: 删除通知
      parameters:
        - name: notificationId
          in: path
          required: true
          schema:
            type: integer
            format: int64

  # ==================== Message ====================
  /orders/{orderId}/messages:
    get:
      tags: [Message]
      summary: 订单聊天记录
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Size"
    post:
      tags: [Message]
      summary: 发送文字消息
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [messageType, content]
              properties:
                messageType:
                  type: string
                  enum: [TEXT]
                content:
                  type: string
                  minLength: 1
                  maxLength: 2000
...
  /orders/{orderId}/reviews:
    get:
      tags: [Review]
      summary: 查看订单评价
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: integer
            format: int64
    post:
      tags: [Review]
      summary: 提交评价
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [rating]
              properties:
                rating:
                  type: integer
                  minimum: 1
                  maximum: 5
                content:
                  type: string
                  maxLength: 500

  # ==================== Report ====================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    Size:
      name: size
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 50
        default: 20
    UserIdPath:
      name: userId
      in: path
      required: true
      schema:
        type: integer
        format: int64
...
  schemas:
...
    UserRole:
      type: string
      enum: [STUDENT, ADMIN]
    UserStatus:
      type: string
      enum: [ACTIVE, DISABLED, ANONYMIZED]
...
    OrderStatus:
      type: string
      enum: [PENDING_CONFIRM, IN_PROGRESS, PENDING_COMPLETION, COMPLETED, CANCELLED, DISPUTE, REVIEWED]
...
    NotificationType:
      type: string
      enum: [APPLICATION, ORDER_STATUS, ORDER_MESSAGE, REVIEW_REQUEST, REPORT_RESULT]

    # ---- Common ----
    ApiResponse:
      type: object
      properties:
        code:
          type: integer
        message:
          type: string
        data:
          type: object
...
    UserProfileResponse:
      allOf:
        - $ref: "#/components/schemas/ApiResponse"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/UserProfile"
    UserProfile:
      type: object
      properties:
        id:
          type: integer
          format: int64
        email:
          type: string
        role:
          $ref: "#/components/schemas/UserRole"
        status:
          $ref: "#/components/schemas/UserStatus"
        verified:
          type: boolean
        profile:
          $ref: "#/components/schemas/ProfileDetail"
        credit:
          $ref: "#/components/schemas/CreditSummary"
        createdAt:
          type: string
          format: date-time
    ProfileDetail:
      type: object
      properties:
        nickname:
          type: string
        avatarUrl:
          type: string
        gender:
          type: string
          enum: [MALE, FEMALE, OTHER]
        grade:
          type: string
        college:
          type: string
        bio:
          type: string
          maxLength: 200
        campus:
          type: string
        contact:
          type: string
        contactVisible:
          type: boolean
    UpdateProfileRequest:
      type: object
      properties:
        nickname:
          type: string
          maxLength: 30
        avatarUrl:
          type: string
        gender:
          type: string
          enum: [MALE, FEMALE, OTHER]
        grade:
          type: string
        college:
          type: string
        bio:
          type: string
          maxLength: 200
        campus:
          type: string
        contact:
          type: string
        contactVisible:
          type: boolean
    PublicProfileResponse:
      allOf:
        - $ref: "#/components/schemas/ApiResponse"
        - type: object
          properties:
            data:
              type: object
              properties:
                userId:
                  type: integer
                  format: int64
                nickname:
                  type: string
                avatarUrl:
                  type: string
                gender:
                  type: string
                college:
                  type: string
                campus:
                  type: string
                verified:
                  type: boolean
                contact:
                  type: string
                  nullable: true
                contactVisible:
                  type: boolean
                creditScore:
                  type: integer
                completedOrders:
                  type: integer
                praiseRate:
                  type: number
                memberSince:
                  type: string
                  format: date
    CreditSummary:
      type: object
      properties:
        score:
          type: integer
        completedOrders:
          type: integer
        praiseRate:
          type: number
    CreditInfoResponse:
      allOf:
        - $ref: "#/components/schemas/ApiResponse"
        - type: object
          properties:
            data:
              type: object
              properties:
                userId:
                  type: integer
                  format: int64
                score:
                  type: integer
                completedOrders:
                  type: integer
                praiseRate:
                  type: number
                recentReviews:
                  type: array
                  items:
                    $ref: "#/components/schemas/ReviewItem"
    ReviewItem:
      type: object
      properties:
        id:
          type: integer
          format: int64
        orderId:
          type: integer
          format: int64
        reviewerNickname:
          type: string
        rating:
          type: integer
        content:
          type: string
        createdAt:
          type: string
          format: date-time

    # ---- Task ----
...